tests/test_organelle.py
# import unittest
# from typing import Dict


# # Mock classes and exceptions to simulate the actual implementations
# class Metabolite:
#     def __init__(self, name, quantity, max_quantity):
#         self.name = name
#         self.quantity = quantity
#         self.max_quantity = max_quantity


# class InsufficientMetaboliteError(Exception):
#     pass


# class QuantityError(Exception):
#     pass


# class UnknownMetaboliteError(Exception):
#     pass


# # Include the Organelle class code (assuming it's defined in the same module for testing purposes)
# class OrganelleMeta(type):
#     _registry = {}

#     def __new__(mcs, name: str, bases: tuple, namespace: dict) -> type:
#         if "name" not in namespace:
#             raise AttributeError(
#                 f"Class '{name}' must define a 'name' class attribute."
#             )
#         cls = super().__new__(mcs, name, bases, namespace)
#         if "structure" not in namespace:
#             setattr(cls, "structure", "Unknown")
#         if name != "Organelle":
#             mcs._registry[name] = cls
#         return cls

#     @classmethod
#     def get_registry(mcs: type) -> dict:
#         return dict(mcs._registry)


# class Organelle(metaclass=OrganelleMeta):
#     name = "Organelle"

#     def __init__(self):
#         self.metabolites: Dict[str, Metabolite] = {}
#         self.add_metabolite("glucose", 0, 1000)
#         self.add_metabolite("atp", 100, 1000)
#         self.add_metabolite("adp", 0, 1000)
#         self.add_metabolite("nad", 10, 1000)
#         self.add_metabolite("nadh", 0, 1000)
#         self.add_metabolite("pyruvate", 0, 1000)
#         self.glycolysis_rate = 1.0

#     def add_metabolite(self, name: str, quantity: int, max_quantity: int) -> None:
#         if not isinstance(name, str):
#             raise TypeError("Metabolite name must be a string.")
#         if not isinstance(quantity, int) or not isinstance(max_quantity, int):
#             raise TypeError("Quantity and max_quantity must be integers.")
#         if quantity < 0 or max_quantity < 0:
#             raise ValueError("Quantity and max_quantity must be non-negative.")
#         if quantity > max_quantity:
#             raise ValueError("Quantity cannot exceed max_quantity.")
#         self.metabolites[name] = Metabolite(name, quantity, max_quantity)

#     def change_metabolite_quantity(self, metabolite_name: str, amount: float) -> None:
#         if not isinstance(metabolite_name, str):
#             raise TypeError("Metabolite name must be a string.")
#         if not isinstance(amount, (int, float)):
#             raise TypeError("Amount must be a number.")
#         if metabolite_name not in self.metabolites:
#             raise UnknownMetaboliteError(f"Unknown metabolite: {metabolite_name}")
#         metabolite = self.metabolites[metabolite_name]
#         new_quantity = metabolite.quantity + amount
#         if new_quantity < 0:
#             raise QuantityError(
#                 f"Cannot reduce {metabolite_name} below zero. Attempted to set {metabolite_name} to {new_quantity}."
#             )
#         if new_quantity > metabolite.max_quantity:
#             raise QuantityError(
#                 f"Cannot exceed max quantity for {metabolite_name}. Attempted to set {metabolite_name} to {new_quantity}, but max is {metabolite.max_quantity}."
#             )
#         metabolite.quantity = new_quantity

#     def is_metabolite_available(self, metabolite: str, amount: float) -> bool:
#         if metabolite not in self.metabolites:
#             raise UnknownMetaboliteError(f"Unknown metabolite: {metabolite}")
#         return self.metabolites[metabolite].quantity >= amount

#     def consume_metabolites(self, **metabolites: float) -> bool:
#         for metabolite, amount in metabolites.items():
#             if not isinstance(metabolite, str):
#                 raise TypeError("Metabolite names must be strings.")
#             if not isinstance(amount, (int, float)):
#                 raise TypeError("Amounts must be numbers.")
#             if amount < 0:
#                 raise QuantityError(
#                     f"Cannot consume a negative amount of {metabolite}."
#                 )
#             if metabolite not in self.metabolites:
#                 raise UnknownMetaboliteError(f"Unknown metabolite: {metabolite}")
#             if self.metabolites[metabolite].quantity < amount:
#                 raise InsufficientMetaboliteError(
#                     f"Insufficient {metabolite} for reaction."
#                 )
#         for metabolite, amount in metabolites.items():
#             self.metabolites[metabolite].quantity -= amount
#         return True

#     def produce_metabolites(self, **metabolites: float) -> bool:
#         for metabolite, amount in metabolites.items():
#             if not isinstance(metabolite, str):
#                 raise TypeError("Metabolite names must be strings.")
#             if not isinstance(amount, (int, float)):
#                 raise TypeError("Amounts must be numbers.")
#             if amount < 0:
#                 raise QuantityError(
#                     f"Cannot produce a negative amount of {metabolite}."
#                 )
#             if metabolite not in self.metabolites:
#                 raise UnknownMetaboliteError(f"Unknown metabolite: {metabolite}")
#             new_quantity = self.metabolites[metabolite].quantity + amount
#             if new_quantity > self.metabolites[metabolite].max_quantity:
#                 raise QuantityError(
#                     f"Cannot exceed max quantity for {metabolite}. Attempted to set {metabolite} to {new_quantity}, but max is {self.metabolites[metabolite].max_quantity}."
#                 )
#         for metabolite, amount in metabolites.items():
#             self.metabolites[metabolite].quantity += amount
#         return True

#     def function(self):
#         raise NotImplementedError("Subclasses must implement the 'function' method.")


# # Unit tests for the Organelle class
# class TestOrganelle(unittest.TestCase):
#     def setUp(self):
#         # Initialize an Organelle instance before each test
#         self.organelle = Organelle()

#     def test_initial_metabolites(self):
#         # Test that initial metabolites are set correctly
#         expected_metabolites = {
#             "glucose": (0, 1000),
#             "atp": (100, 1000),
#             "adp": (0, 1000),
#             "nad": (10, 1000),
#             "nadh": (0, 1000),
#             "pyruvate": (0, 1000),
#         }
#         for name, (quantity, max_quantity) in expected_metabolites.items():
#             self.assertIn(name, self.organelle.metabolites)
#             metabolite = self.organelle.metabolites[name]
#             self.assertEqual(metabolite.quantity, quantity)
#             self.assertEqual(metabolite.max_quantity, max_quantity)

#     def test_add_metabolite(self):
#         # Test adding a valid metabolite
#         self.organelle.add_metabolite("oxygen", 50, 100)
#         self.assertIn("oxygen", self.organelle.metabolites)
#         metabolite = self.organelle.metabolites["oxygen"]
#         self.assertEqual(metabolite.quantity, 50)
#         self.assertEqual(metabolite.max_quantity, 100)

#         # Test invalid metabolite name type
#         with self.assertRaises(TypeError):
#             self.organelle.add_metabolite(123, 50, 100)

#         # Test invalid quantity types
#         with self.assertRaises(TypeError):
#             self.organelle.add_metabolite("oxygen", "50", 100)
#         with self.assertRaises(TypeError):
#             self.organelle.add_metabolite("oxygen", 50, "100")

#         # Test negative quantities
#         with self.assertRaises(ValueError):
#             self.organelle.add_metabolite("oxygen", -10, 100)
#         with self.assertRaises(ValueError):
#             self.organelle.add_metabolite("oxygen", 10, -100)

#         # Test quantity exceeding max_quantity
#         with self.assertRaises(ValueError):
#             self.organelle.add_metabolite("oxygen", 150, 100)

#     def test_change_metabolite_quantity(self):
#         # Test increasing quantity within limits
#         self.organelle.change_metabolite_quantity("atp", 50)
#         self.assertEqual(self.organelle.metabolites["atp"].quantity, 150)

#         # Test decreasing quantity within limits
#         self.organelle.change_metabolite_quantity("atp", -100)
#         self.assertEqual(self.organelle.metabolites["atp"].quantity, 50)

#         # Test exceeding max_quantity
#         with self.assertRaises(QuantityError):
#             self.organelle.change_metabolite_quantity("atp", 1000)

#         # Test reducing quantity below zero
#         with self.assertRaises(QuantityError):
#             self.organelle.change_metabolite_quantity("atp", -200)

#         # Test unknown metabolite
#         with self.assertRaises(UnknownMetaboliteError):
#             self.organelle.change_metabolite_quantity("unknown", 10)

#         # Test invalid types
#         with self.assertRaises(TypeError):
#             self.organelle.change_metabolite_quantity(123, 10)
#         with self.assertRaises(TypeError):
#             self.organelle.change_metabolite_quantity("atp", "10")

#     def test_is_metabolite_available(self):
#         # Test availability when sufficient
#         self.assertTrue(self.organelle.is_metabolite_available("nad", 5))

#         # Test availability when insufficient
#         self.assertFalse(self.organelle.is_metabolite_available("nad", 20))

#         # Test unknown metabolite
#         with self.assertRaises(UnknownMetaboliteError):
#             self.organelle.is_metabolite_available("unknown", 5)

#     def test_consume_metabolites(self):
#         # Test consuming available metabolites
#         result = self.organelle.consume_metabolites(atp=50, nad=5)
#         self.assertTrue(result)
#         self.assertEqual(self.organelle.metabolites["atp"].quantity, 50)
#         self.assertEqual(self.organelle.metabolites["nad"].quantity, 5)

#         # Test consuming more than available
#         with self.assertRaises(InsufficientMetaboliteError):
#             self.organelle.consume_metabolites(atp=100)

#         # Test consuming unknown metabolite
#         with self.assertRaises(UnknownMetaboliteError):
#             self.organelle.consume_metabolites(unknown=10)

#         # Test invalid metabolite name type
#         with self.assertRaises(TypeError):
#             self.organelle.consume_metabolites(**{123: 10})

#         # Test invalid amount type
#         with self.assertRaises(TypeError):
#             self.organelle.consume_metabolites(atp="10")

#         # Test negative consumption amount
#         with self.assertRaises(QuantityError):
#             self.organelle.consume_metabolites(atp=-10)

#     def test_produce_metabolites(self):
#         # Test producing metabolites within limits
#         result = self.organelle.produce_metabolites(pyruvate=500)
#         self.assertTrue(result)
#         self.assertEqual(self.organelle.metabolites["pyruvate"].quantity, 500)

#         # Test exceeding max_quantity
#         with self.assertRaises(QuantityError):
#             self.organelle.produce_metabolites(pyruvate=600)

#         # Test producing unknown metabolite
#         with self.assertRaises(UnknownMetaboliteError):
#             self.organelle.produce_metabolites(unknown=10)

#         # Test invalid metabolite name type
#         with self.assertRaises(TypeError):
#             self.organelle.produce_metabolites(**{123: 10})

#         # Test invalid amount type
#         with self.assertRaises(TypeError):
#             self.organelle.produce_metabolites(pyruvate="10")

#         # Test negative production amount
#         with self.assertRaises(QuantityError):
#             self.organelle.produce_metabolites(pyruvate=-10)

#     def test_function_not_implemented(self):
#         # Test that 'function' method raises NotImplementedError
#         with self.assertRaises(NotImplementedError):
#             self.organelle.function()

#     def test_metaclass_registry(self):
#         # Test that OrganelleMeta correctly registers subclasses
#         class Mitochondria(Organelle):
#             name = "Mitochondria"

#             def function(self):
#                 pass

#         registry = OrganelleMeta.get_registry()
#         self.assertIn("Mitochondria", registry)
#         self.assertIs(registry["Mitochondria"], Mitochondria)

#     def test_missing_name_attribute(self):
#         # Test that missing 'name' attribute raises AttributeError
#         with self.assertRaises(AttributeError):

#             class InvalidOrganelle(Organelle):
#                 pass  # 'name' attribute is missing

#     # If 'function' method enforcement is enabled in the metaclass, you can test it as follows:
#     # def test_missing_function_method(self):
#     #     with self.assertRaises(NotImplementedError):
#     #         class IncompleteOrganelle(Organelle):
#     #             name = "IncompleteOrganelle"


# # Run the tests
# if __name__ == "__main__":
#     unittest.main()
